+++ src/components/manage/Widgets/ArrayWidget.jsx
@@ -104,9 +104,9 @@ const customSelectStyles = {
   }),
 };
 
-const getChoices = choices =>
+/*const getChoices = choices =>
   choices ? choices.map(item => ({ label: item, value: item })) : [];
-
+*/
 /**
  * ArrayWidget component class.
  * @class ArrayWidget
@@ -222,7 +222,7 @@ class ArrayWidget extends Component {
     this.props.getVocabulary(this.vocabBaseUrl, search, offset);
     this.setState({ search });
     return {
-      options: this.props.choices,
+      options: this.props.items.choices,
       hasMore: this.props.itemsTotal > 25,
       additional: {
         offset: offset === additional.offset ? offset + 25 : offset,
@@ -270,7 +270,7 @@ class ArrayWidget extends Component {
                 <AsyncPaginate
                   className="react-select-container"
                   classNamePrefix="react-select"
-                  options={this.props.choices || []}
+                  options={this.props.items.choices || []}
                   styles={customSelectStyles}
                   theme={selectTheme}
                   components={{ DropdownIndicator, Option }}
@@ -287,7 +287,15 @@ class ArrayWidget extends Component {
                   className="react-select-container"
                   classNamePrefix="react-select"
                   options={
-                    this.props.choices || getChoices(this.props.default) || []
+                    this.props.items.choices
+                      ? [
+                          ...this.props.items.choices.map(option => ({
+                            value: option[0],
+                            label: option[1],
+                          })),
+                          { label: 'No value', value: 'no-value' },
+                        ]
+                      : []
                   }
                   styles={customSelectStyles}
                   theme={selectTheme}